# PERF: Sparse ops speedup