@jreback 2016-05-05
thanks!

and hopefully turned off annoying codecov warnings.